Bugfix for dynamic dropdown result processing
Fixed two bugs:

The dynamic dropdown was hardcoded to only handle the case where the selection
items were in the first position of the return array and were at the first level
of the object at that position.
Now the position of the object in the return array and the path into the object
to the selection items can be specified.

Some dynamic service functions don't like empty input, but the DD will always
query the server regarless of the input contents. This behavior can now
be configured.

These changes require https://github.com/kbase/narrative_method_store/pull/74

There are no meaningful tests for the DD, and developing  a test harness
for the DD is way out of scope, so the changes were tested manually.