Incorporated feedback on app cell info tab.
This commit adjusts the app cell info tab according to user feedback, fixes
some unit tests and updates some linter settings. Note that this commit changes
the order of the info tab among other app cell tabs. This is not immediately
obvious as JS did not always guarantee key order in objects, but now it does.
It removes some instances of Bluebird in favor of the global built in Promise
object. It fixes some unit tests to test for what their spec says they are
testing for, and includes some new tests.