Debug issue with narrative_version
The narrative_version that was built and pushed by GHA has an old version number 4.2.0 for some reason, might be an issue with when the feature branch was branched off and merged back in versus when the most recent merge of develop into master occurred. Debugging it with a print during the build.